# Conflicts:
#	backend/app/api/v1/api.py
#	backend/app/api/v1/auth/auth.py
#	backend/app/api/v1/casbin.py
#	backend/app/api/v1/dept.py
#	backend/app/api/v1/dict_data.py
#	backend/app/api/v1/menu.py
#	backend/app/api/v1/role.py
#	backend/app/api/v1/task.py
#	backend/app/api/v1/user.py
#	backend/app/services/api_service.py
#	backend/app/services/casbin_service.py
#	backend/app/services/dept_service.py
#	backend/app/services/dict_data_service.py
#	backend/app/services/dict_type_service.py
#	backend/app/services/menu_service.py
#	backend/app/services/opera_log_service.py
#	backend/app/services/role_service.py
#	backend/app/services/task_service.py
#	backend/app/services/user_service.py

+++ backend/app/services/menu_service.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
 from typing import Any
 
 from fastapi import Request
@@ -12,107 +11,28 @@ from backend.app.crud.crud_menu import MenuDao
 from backend.app.crud.crud_role import RoleDao
 from backend.app.database.db_mysql import async_db_session
 from backend.app.models import Menu
-from backend.app.schemas.menu import CreateMenu, UpdateMenu
+from backend.app.schemas.menu import CreateMenuParam, UpdateMenuParam
 from backend.app.utils.build_tree import get_tree_data
 
 
-class MenuServiceABC(ABC):
-    """
-    菜单服务基类
-    """
-
-    @abstractmethod
-    async def get(self, *, pk: int) -> Menu:
-        """
-        获取菜单详情
-
-        :param pk:
-        :return:
-        """
-        pass
-
-    @abstractmethod
-    async def get_menu_tree(self, *, title: str | None = None, status: int | None = None) -> list[dict[str, Any]]:
-        """
-        获取所有菜单展示树
-
-        :param title:
-        :param status:
-        :return:
-        """
-        pass
-
-    @abstractmethod
-    async def get_role_menu_tree(self, *, pk: int) -> list[dict[str, Any]]:
-        """
-        获取角色所有菜单
-
-        :param pk:
-        :return:
-        """
-        pass
-
-    @abstractmethod
-    async def get_user_menu_tree(self, *, request: Request) -> list[dict[str, Any]]:
-        """
-        获取用户菜单展示树
-
-        :param request:
-        :return:
-        """
-        pass
-
-    @abstractmethod
-    async def create(self, *, obj: CreateMenu) -> None:
-        """
-        创建菜单
-
-        :param obj:
-        :return:
-        """
-        pass
-
-    @abstractmethod
-    async def update(self, *, pk: int, obj: UpdateMenu) -> int:
-        """
-        更新菜单
-
-        :param pk:
-        :param obj:
-        :return:
-        """
-        pass
-
-    @abstractmethod
-    async def delete(self, *, pk: int) -> int:
-        """
-        删除菜单
-
-        :param pk:
-        :return:
-        """
-        pass
-
-
-class MenuServiceImpl(MenuServiceABC):
-    """
-    菜单服务实现类
-    """
-
-    async def get(self, *, pk: int) -> Menu:
+class MenuService:
+    @staticmethod
+    async def get(*, pk: int) -> Menu:
         async with async_db_session() as db:
             menu = await MenuDao.get(db, menu_id=pk)
             if not menu:
                 raise errors.NotFoundError(msg='菜单不存在')
             return menu
 
-    async def get_menu_tree(self, *, title: str | None = None, status: int | None = None) -> list[dict[str, Any]]:
+    @staticmethod
+    async def get_menu_tree(*, title: str | None = None, status: int | None = None) -> list[dict[str, Any]]:
         async with async_db_session() as db:
             menu_select = await MenuDao.get_all(db, title=title, status=status)
             menu_tree = await get_tree_data(menu_select)
             return menu_tree
 
-    async def get_role_menu_tree(self, *, pk: int) -> list[dict[str, Any]]:
+    @staticmethod
+    async def get_role_menu_tree(*, pk: int) -> list[dict[str, Any]]:
         async with async_db_session() as db:
             role = await RoleDao.get_with_relation(db, pk)
             if not role:
@@ -122,7 +42,8 @@ class MenuServiceImpl(MenuServiceABC):
             menu_tree = await get_tree_data(menu_select)
             return menu_tree
 
-    async def get_user_menu_tree(self, *, request: Request) -> list[dict[str, Any]]:
+    @staticmethod
+    async def get_user_menu_tree(*, request: Request) -> list[dict[str, Any]]:
         async with async_db_session() as db:
             roles = request.user.roles
             menu_ids = []
@@ -134,7 +55,8 @@ class MenuServiceImpl(MenuServiceABC):
                 menu_tree = await get_tree_data(menu_select)
             return menu_tree
 
-    async def create(self, *, obj: CreateMenu) -> None:
+    @staticmethod
+    async def create(*, obj: CreateMenuParam) -> None:
         async with async_db_session.begin() as db:
             title = await MenuDao.get_by_title(db, obj.title)
             if title:
@@ -145,7 +67,8 @@ class MenuServiceImpl(MenuServiceABC):
                     raise errors.NotFoundError(msg='父级菜单不存在')
             await MenuDao.create(db, obj)
 
-    async def update(self, *, pk: int, obj: UpdateMenu) -> int:
+    @staticmethod
+    async def update(*, pk: int, obj: UpdateMenuParam) -> int:
         async with async_db_session.begin() as db:
             menu = await MenuDao.get(db, pk)
             if not menu:
@@ -157,17 +80,17 @@ class MenuServiceImpl(MenuServiceABC):
                 parent_menu = await MenuDao.get(db, obj.parent_id)
                 if not parent_menu:
                     raise errors.NotFoundError(msg='父级菜单不存在')
+            if obj.parent_id == menu.id:
+                raise errors.ForbiddenError(msg='禁止关联自身为父级')
             count = await MenuDao.update(db, pk, obj)
             await redis_client.delete_prefix(settings.PERMISSION_REDIS_PREFIX)
             return count
 
-    async def delete(self, *, pk: int) -> int:
+    @staticmethod
+    async def delete(*, pk: int) -> int:
         async with async_db_session.begin() as db:
             children = await MenuDao.get_children(db, pk)
             if children:
                 raise errors.ForbiddenError(msg='菜单下存在子菜单，无法删除')
             count = await MenuDao.delete(db, pk)
             return count
-
-
-MenuService: MenuServiceABC = MenuServiceImpl()
